requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Collection
requests>=2.31.0
eia-python>=1.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0
geopandas>=0.13.0
shapely>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.4
xgboost>=2.0.0
pmdarima>=2.0.0

# Interactive Dashboard
streamlit>=1.24.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Utilities
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.23.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data validation
great-expectations>=0.18.0

